Floating Healthbar
This is a health floating bar. It can be used to 2D game with few adjuments but for 3D games i think must be put a little more effort because of the z axis. When the player get far the hearts stay at the same size. Try to calculate it using the ditance to the camera e resizing it.